# openapi3.yaml
openapi: 3.0.1
info:
  title: metatile-queue-populator
  description: Service for adding metatiles to a queue based on variety of inputs
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

paths:
  /tiles/list:
    post:
      operationId: postTilesList
      summary: Add metatiles to the queue
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TilesListRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    description: message
        400:
          $ref: '#/components/responses/BadRequest'
        5XX:
          $ref: '#/components/responses/UnexpectedError'

  /tiles/bbox:
    post:
      operationId: postTilesByBbox
      summary: add the all the meta tiles in the given bbox and zoom range into the queue
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BboxTilesRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    description: message
        400:
          $ref: '#/components/responses/BadRequest'
        409:
          $ref: '#/components/responses/Conflict'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnexpectedError:
      description: Unexpected Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string

    TilesListRequest:
      type: array
      minItems: 1
      items:
        type: object
        required:
          - z
          - x
          - y
        properties:
          x:
            type: integer
          y:
            type: integer
          z:
            type: integer
          metatile:
            type: integer
            default: 1
    BboxTilesRequest:
      type: object
      required:
        - bbox
        - minZoom
        - maxZoom
      properties:
        bbox:
          type: array
          description: WGS84 bbox
          items:
            type: number
          minItems: 4
          maxItems: 4
          example: [-90, -90, 90, 90]
        minZoom:
          type: integer
          minimum: 0
          maximum: 18
          description: The minimum zoom for which metatiles will be calculated
        maxZoom:
          type: integer
          minimum: 0
          maximum: 18
          description: The maximum zoom for which metatiles will be calculated. note - the value should be higher or equal than minZoom